sql: introduce UUID field type
This patch introduces UUID to SQL. UUID is now available as a new field
type.

Closes #5886

@TarantoolBot document
Title: Field type UUID is now available in SQL

The UUID field type is now available in SQL. This means that we can
create spaces and indexes with UUID, use it in SELECT, UPDATE and
DELETE. UUID can be accepted and returned by built-in functions and
user-defined functions.

During comparison, UUID cannot be implicitly converted to any other
types, but in any other case, UUID can be implicitly converted to STRING
or VARBINARY. UUID can be explicitly converted to STRING or VARBINARY
using CAST(). STRING and VARBINARY can be implicitly converted to UUID,
except for comparison. STRING and VARBINARY can be explicitly converted
to UUID using CAST().