txn_limbo: move lsn assignment to journal completion callback
Previously local lsn assignment for txn_limbo entries was done right in
txn_commit() body after the WAL write.

In order to make the entries lsn available to on_wal_write triggers, which are
called from the journal completion callback, assign it right in the
callback before the triggers are called.

Note, ACKing the entry is not moved to journal completion callback, since
it may trigger writing CONFIRM, which cannot be done from the sched fiber
(which executes the journal callback).

Prerequisite #6032