sql: refactor LLVM JIT, change base LLVM module generation
Transit from LLVM Orcv1 to Orcv2 C API and refactor the whole infrastructure.

Minor cosmetical and logical code changes due to the mentioned refactoring.

Instead of generating only types that we use and references to externally called functions, generate implementations of functions we use, so that we can inline them.