sql: expose VDBE interface and introduce callbacks emulating VDBE opcodes
Expose vdbe_field_ref_fetch so that it can be used in LLVM JIT compilation of TK_COLUMN_REF's. Add callbacks that can be used to emulate various VDBE opcodes in LLVM JIT compilation: in the future, these are actually going to be inlined. We do this to rely on clang's LLVM IR generation rather then manually generating IR.