memtx: add assertion and test for gap tracker retention during rollback
Gap trackers are stored in the story at the top of the history chain (see
`memtx_tx_story_{un}link_top_light`): following this logic, during rollback
we need to rebind them to the older story. We already maintain this in
8c24259, since a story is not retained iff it has a newer story (which
means it does not store trackers) or if it has an older story (in which
case rebinding is done in `memtx_tx_story_{un}link_top_light`): add an
assertion that a story deleted during rollback does not have any gap
trackers and a test for this case.

Follow-up 8c24259

NO_CHANGELOG=internal
NO_DOC=internal