memtx: support multikey indexes in MVCC transaction manager
Multikey indexes require separate history chains for each of the keys: this
is required for tuple clarification as well as for gap tracking. In order
to achieve this, we add another layer of indirection via a key array for
each story: history chains are now formed by keys rather than stories.

Closes #6385

NO_DOC=<the documentation doesn't mention that multikey indexes are not
        supported by memtx MVCC transaction manager>