net.box: separately handle `box.error` in `handle_eval_result`
Currently, in the error case, `handle_eval_result` constructs a new error
passing the error it receives as an argument (which can be a box.error) to
the message argument of the `box.error` constructor.

Then, during `box.error` construction, if the error argument of
`handle_eval_result` was a `box.error`, it is converted to a string using
its string conversion metamethod, which currently returns the error's
message, losing the rest of the error's information.

But in scope of #9105, we are going to increase the verbosity of the
string conversion metamethod, and it will not return the error's message.
To adjust to this new behaviour, let's separately handle `box.error` in
`handle_eval_result`, and create the same error, but with the `ER_PROC_LUA`
code.

Needed for #9105

NO_CHANGELOG=<internal change>
NO_TEST=<internal change>
NO_DOC=<internal change>