replication: infinite `replication_synchro_timeout`
This commit allows the `replication_synchro_timeout` option to be set to
0, which means an infinite timeout. When set to 0, the transaction will
never be rolled back due to timeout, only PROMOTE and DEMOTE can rollback
it in this case. This behavior is more consistent with the standard Raft
design.

Part of #7486

@TarantoolBot document
Title: new possible value for 'replication_synchro_timeout': 0 (infinity)
Product: Tarantool
Since: 3.2

Root document: https://www.tarantool.io/ru/doc/latest/reference/configuration/#cfg-replication-replication-synchro-timeout

Now it is allowed to set the option value to 0, which means an infinite
timeout (the transaction will not be rolled back due to a timeout).

NO_CHANGELOG=Will be added in another commit