memtx: factor out multikey loop from regular and functional `replace`
Currently, memtx multikey index variants encapsulate the processing of
multiple multikey keys. This is bad for to reasons.

First, this violates the memtx index `replace` interface. Each `replace`
operation is supposed to return a `replaced` and a `successor` tuple for
tracking a transaction's read and write sets in memtx MVCC. However,
multikey index variants yield multiple `replaced` and `successor` tuples
which cannot be returned through the existing interface.

Second, this makes the replace operation coarse grained. Effectively, we
lose control of the index state of each multikey key: either all the keys
are inserted the index or they are deleted from it. However, in the context
of memtx MVCC, each multikey key has an independent history chain, and thus
its presence in the index needs to managed independently.

Together, these discrepancies prevent support of multikey index variants in
memtx MVCC (see also  #12313).

Last but not least, this complicates the indexing logic. The `replace`
operation for multikey index variants builds complex insertion, deletion
and rolling back loops on top of the regular index replace operation. This
is bad, as the `replace` operation should be as simple as possible and it
should not be responsible for multiple operations.

To achieve this, we factor out the multikey loop from the multikey and
functional index replace variants as into `memtx_index_replace`.
Internally, we change the interface of `replace` to use a new abstraction,
`memtx_index_entry`, that generalizes the tuple abstraction for multikey
and functional indexes by passing the tuple together with its multikey
array index or functional key to uniquely identify such entries. We also
pull out excluding logic from all `replace` variants.

Closes #12386
Needed for #6385

NO_CHANGELOG=<refactoring>
NO_DOC=<refactoring>
NO_TEST=<refactoring>